`ViewerComicSpinType`, `ViewerSpinType`도 demo viewer에서 직접 관리하도록 함
