+++ demo/src/components/settings/NovelSpineSetting.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { selectReaderSetting, ViewerSpinType } from '../../../../lib';
+import { selectReaderSetting } from '../../../../lib';
+import { ViewerSpinType } from '../../constants/SettingConstants';
 import SpinButton from './SpinButton';
 import SvgIcons from '../icons/SvgIcons';
