+++ demo/src/constants/SettingConstants.js
@@ -1,0 +1,56 @@
+import makeConstants from '../../../src/util/Constant';
+import { updateObject } from '../../../src/util/Util';
+
+const _ViewerSpinType = {
+  FONT_SIZE: 'font_size',
+  PADDING: 'padding',
+  LINE_HEIGHT: 'line_height',
+};
+
+export const ViewerSpinType = makeConstants(updateObject(_ViewerSpinType, {
+  _LIST: [
+    _ViewerSpinType.FONT_SIZE,
+    _ViewerSpinType.PADDING,
+    _ViewerSpinType.LINE_HEIGHT,
+  ],
+  _STRING_MAP: {
+    [_ViewerSpinType.FONT_SIZE]: '글자 크기',
+    [_ViewerSpinType.PADDING]: '문단 너비',
+    [_ViewerSpinType.LINE_HEIGHT]: '줄 간격',
+  },
+}), {
+  toReaderSettingType: (type) => {
+    switch (type) {
+      case _ViewerSpinType.FONT_SIZE:
+        return 'fontSizeLevel';
+      case _ViewerSpinType.PADDING:
+        return 'paddingLevel';
+      case _ViewerSpinType.LINE_HEIGHT:
+        return 'lineHeightLevel';
+      default:
+        return 'fontSizeLevel';
+    }
+  },
+});
+
+const _ViewerComicSpinType = {
+  CONTENT_WIDTH: 'width',
+};
+
+export const ViewerComicSpinType = makeConstants(updateObject(_ViewerComicSpinType, {
+  _LIST: [
+    _ViewerComicSpinType.CONTENT_WIDTH,
+  ],
+  _STRING_MAP: {
+    [_ViewerComicSpinType.CONTENT_WIDTH]: '콘텐츠 너비',
+  },
+}), {
+  toReaderSettingType: (type) => {
+    switch (type) {
+      case _ViewerComicSpinType.CONTENT_WIDTH:
+        return 'contentWidthLevel';
+      default:
+        return 'contentWidthLevel';
+    }
+  },
+});
